### 2021-09-29 #3778 @christinahedges
Hi @richteague, thanks for your submission to JOSS! I took a quick look at your submission and I have these comments:

### Tests

I don't believe there are tests in your repo. One of the review checklist criteria for JOSS is:
```
Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
```
You can see the checklist [here](https://joss.readthedocs.io/en/latest/review_checklist.html). Before I find reviewers for this submission, do you want some time to address this checklist item?

### Licensing

I see that you have used the `detect_peaks.py` module from [this](https://joss.readthedocs.io/en/latest/review_checklist.html) repo. I believe you need to make some changes to this file to be compliant with their [license](https://github.com/BMClab/BMC/blob/v0.0.2/LICENSE.txt) by adding their license and copyright notice at the top of the file, and stating any changes.

### Reviewers

As I start to look for reviewers for this submission, I wanted to ask is there anyone you can identify you believe would be a good reviewer for this work? They would need to have no [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
