### 2021-09-30 331574c
included license for detect_peaks
