+++ disksurf/detect_peaks.py
@@ -1,8 +1,42 @@
+"""
+MIT License
+
+Copyright (c) 2020 Marcos Duarte
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+Detect peaks in data based on their amplitude and other features.
+
+Originally written by Marcos Duarte with an MIT License.
+Only modification by Richard Teague was removal of warnings.
+"""
+
+from __future__ import division, print_function
 import numpy as np
+
+__author__ = "Marcos Duarte, https://github.com/demotu"
+__version__ = "1.0.6"
+__license__ = "MIT"
 
 
 def detect_peaks(x, mph=None, mpd=1, threshold=0, edge='rising',
-                 kpsh=False, valley=False, show=False):
+                 kpsh=False, valley=False, show=False, ax=None, title=True):
 
     """Detect peaks in data based on their amplitude and other features.
     Parameters
@@ -10,7 +44,9 @@ def detect_peaks(x, mph=None, mpd=1, threshold=0, edge='rising',
     x : 1D array_like
         data.
     mph : {None, number}, optional (default = None)
-        detect peaks that are greater than minimum peak height.
+        detect peaks that are greater than minimum peak height (if parameter
+        `valley` is False) or peaks that are smaller than maximum peak height
+         (if parameter `valley` is True).
     mpd : positive integer, optional (default = 1)
         detect peaks that are at least separated by minimum peak distance (in
         number of data).
@@ -27,6 +63,10 @@ def detect_peaks(x, mph=None, mpd=1, threshold=0, edge='rising',
         if True (1), detect valleys (local minima) instead of peaks.
     show : bool, optional (default = False)
         if True (1), plot data in matplotlib figure.
+    ax : a matplotlib.axes.Axes instance, optional (default = None).
+    title : bool or string, optional (default = True)
+        if True, show standard title. If False or empty string, doesn't show
+        any title. If string, shows string as title.
     Returns
     -------
     ind : 1D array_like
@@ -39,8 +79,41 @@ def detect_peaks(x, mph=None, mpd=1, threshold=0, edge='rising',
     See this IPython Notebook [1]_.
     References
     ----------
-    .. [1] http://nbviewer.ipython.org/github/demotu/BMC/
[…]
