### 2021-09-30 #3778 @richteague
Thanks @christinahedges!

I've updated the license for `detect_peaks.py` with the MIT version that it was originally written under.

Regarding the tests, we've developed a [tutorial notebook](https://disksurf.readthedocs.io/en/latest/tutorials/tutorial_1.html) which users can run that should act as a test that everything is installed correctly. It this sufficient for the "manual steps" described in the checklist criteria, or should we make a separate testing suite?

I will have a think about potential reviewers while we resolve the testing issue.
