### 2021-09-30 #3778 @christinahedges
Hi @richteague! Thanks for updating to be compliant with the license.

Re tests, these are the instructions from JOSS:
```
Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

* Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
* OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
* Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works
```
Based on this, I would say that having a tutorial notebook for testing comes under the "OK" bracket. I think because it meets this standard, I'm happy to let the compliance with testing (e.g. test coverage) be discussed with reviewers.

Let me know if you have any reviewer suggestions, I will also look to find appropriate reviewers.
