### 2021-10-12 8b1fdf1
typo in MCMC fitting
