### 2021-11-01 #3827 @emptymalei
Hmm, I found a strange thing about the markdown to pdf generator.

![CleanShot 2021-11-01 at 20 11 44@2x](https://user-images.githubusercontent.com/663798/139727777-25159df9-d72a-4f25-b854-c1bdabb5282a.png)

The order of `'` and `,` is different from the original markdown file:

https://github.com/richteague/disksurf/blob/8b1fdf1c1780eefa6c6d008ca4a6739a5a1f45df/paper/paper.md?plain=1#L40
