### 2021-11-01 #3827 @richteague
Apologies for missing that code, @cpinte! Do you have a Zenodo reference that I could use to reference it, or would a link to the GitHub repository be OK?

@kthyng - Do you have a preference from an editorial point of view?
