### 2021-11-01 richteague/disksurf#4 @emptymalei
__future__ in detect_peaks
Thanks for the nice package. 🎉

I am wondering if the following imports are really needed, as this package already requires python 3. 🤔

https://github.com/richteague/disksurf/blob/8b1fdf1c1780eefa6c6d008ca4a6739a5a1f45df/disksurf/detect_peaks.py#L30

(https://github.com/openjournals/joss-reviews/issues/3827)
