### 2021-11-01 #3827 @danielskatz
It is odd that it's being generated differently than specified, though in this case, correct grammar would be

> 'scattering surface,' the emission surface equivalent for small, submicron grains, was described

both the first comma inside the quote, and the second comma added.
