### 2021-11-01 #3827 @kthyng
I didn't know that would happen in the pdf compilation! My understanding for the order of quotes vs. commas in a sentence is geographically dependent with different accepted rules in the U.S. vs the U.K. for example.
