+++ paper/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 The Atacama Millimeter/submillimeter Array (ALMA) has brought our view of protoplanetary disks, the formation environment of planets, into sharp focus. The unparalleled angular resolution now achievable with ALMA allows us to routinely resolve the 3D structure of these disks; detailing the vertical structure of the gas and dust from which planets are formed. Extracting the precise height from where emission arises is a key step towards understanding the conditions in which a planet is born, and, in turn, how the planet can affect the parental disk.
 
-A method for extracting a 'scattering surface', the emission surface equivalent for small, submicron grains was described in @Stolker:2016 who provided the `diskmap` package. However, this approach is not suitable for molecular emission, which traces the gas component of the disk and has a strong frequency dependence due to Doppler shifting from the disk rotation. @Pinte:2018 presented an alternative method that could account for this frequency dependence and demonstrated that this could be used to trace key physical properties of the protoplanetary disk, namely the gas temperature and rotation velocity, along the emission surface.
+A method for extracting a 'scattering surface', the emission surface equivalent for small, submicron grains was described in @Stolker:2016 who provided the `diskmap` package. However, this approach is not suitable for molecular emission, which traces the gas component of the disk and has a strong frequency dependence due to Doppler shifting from the disk rotation. @Pinte:2018 presented an alternative method that could account for this frequency dependence and demonstrated that this could be used to trace key physical properties of the protoplanetary disk, namely the gas temperature and rotation velocity, along the emission surface. An example script was provided with this publication demonstrating the algorithm: https://github.com/cpinte/CO_layers.
 
 While the measurement of the emission surface only requires simple geometrical transformations, the largest source of uncertainty arises through the handling of noisy data. As more works perform such analyses, for example @Teague:2019, @Rich:2021, and @Law:2021, the need for an easy-to-use package that implements this method was clear. Such a package would facilitate the rapid reproduction of published results, enable direct comparisons between numerical simulations and observations [@Calahan:2021; @Schwarz:2021], and ease benchmarking between different publications. `disksurf` provides this functionality, along with a tutorial to guide users through the process of extracting an emission surface. The code is developed in such a way that as the quality of observations improve, the extraction methods can be easily refined to maintain precise measurements of the emission surface.
 
