### 2021-11-08 richteague/disksurf#4 @richteague
Good catch, I've removed this with https://github.com/richteague/disksurf/commit/451923a1d2dce9fb51f57fdf4696b2b5120cdb6e.
