+++ paper/paper.bib
@@ -94,19 +94,17 @@ archivePrefix = {arXiv},
 }
 
 @article{Law:2021,
-       author = {{Law}, Charles J. and {Teague}, Richard and {Loomis}, Ryan A. and {Bae}, Jaehan and {{\"O}berg}, Karin I. and {Czekala}, Ian and {Andrews}, Sean M. and {Aikawa}, Yuri and {Alarc{\'o}n}, Felipe and {Bergin}, Edwin A. and {Bergner}, Jennifer B. and {Booth}, Alice S. and {Bosman}, Arthur D. and {Calahan}, Jenny K. and {Cataldi}, Gianni and {Cleeves}, L. Ilsedore and {Furuya}, Kenji and {Guzm{\'a}n}, Viviana V. and {Huang}, Jane and {Ilee}, John D. and {Le Gal}, Romane and {Liu}, Yao and {Long}, Feng and {M{\'e}nard}, Fran{\c{c}}ois and {Nomura}, Hideko and {P{\'e}rez}, Laura M. and {Qi}, Chunhua and {Schwarz}, Kamber R. and {Soto}, Daniela and {Tsukagoshi}, Takashi and {Yamato}, Yoshihide and {van't Hoff}, Merel L.~R. and {Walsh}, Catherine and {Wilner}, David J. and {Zhang}, Ke},
-        title = "{Molecules with ALMA at Planet-forming Scales (MAPS) IV: Emission Surfaces and Vertical Distribution of Molecules}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
-         year = 2021,
-        month = sep,
-          eid = {arXiv:2109.06217},
-        pages = {arXiv:2109.06217},
-archivePrefix = {arXiv},
-       eprint = {2109.06217},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021arXiv210906217L},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    doi = {10.3847/1538-4365/ac1434},
+    url = {https://doi.org/10.3847/1538-4365/ac1434},
+    year = 2021,
+    month = {nov},
+    publisher = {American Astronomical Society},
+    volume = {257},
+    number = {1},
+    pages = {3},
+    author = {Charles J. Law and Ryan A. Loomis and Richard Teague and Karin I. Öberg and Ian Czekala and Sean M. Andrews and Jane Huang and Yuri Aikawa and Felipe Alarc{\'{o}}n and Jaehan Bae and Edwin A. Bergin and Jennifer B. Bergner and Yann Boehler and Alice S. Booth and Arthur D. Bosman and Jenny K. Calahan and Gianni Cataldi and L. Ilsedore Cleeves and Kenji Furuya and Viviana V. Guzm{\'{a}}n and John D. Ilee and Romane Le Gal and Yao Liu and Feng Long and Fran{\c{c}}ois M{\'{e}}nard and Hideko Nomura and Chunhua Qi and Kamber R. Schwarz and Anibal Sierra and Takashi Tsukagoshi and Yoshihide Yamato and Merel L. R. van 't Hoff and Catherine Walsh and David J. Wilner and Ke Zhang},
+    title = {Molecules with {ALMA} at Planet-forming Scales ({MAPS}). {III}. Characteristics of Radial Chemical Substructures},
[…]
