### 2021-11-10 #3827 @christinahedges
Thank you very much reviewers! @richteague your paper submission looks good to me. If you are happy and would like to make no further changes based on the review, please make a tagged release 🏷️ of the code and archive it, and then please report the version number and archive DOI here in this thread. Please make sure this archived version has all the correct meta data (author names, title etc)!

I can then recommend to our EIC that we accept this submission.

Thank you again to @emptymalei and @cpinte for your hard work and thorough review!
