### 2021-11-10 #3827 @richteague
Thanks @christinahedges! I've made a version of the code and archived version here: 10.5281/zenodo.5670586.

Thanks also to @cpinte and @emptymalei for reviewing this submission!
