### 2021-11-12 #3827 @kyleniemeyer
Hi @richteague, can you edit the metadata of the Zenodo archive so that the title and author list match the JOSS paper?
